ml-service/app/requirements.txt
fastapi
uvicorn[standard]
torch
numpy
librosa
soundfile
matplotlib
kafka-python
python-dotenv